feat: INFRA-567 Added custom parameters. Changed repo module to point to refer to repo name. Moved generalizable templating code to separate module
