+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "ferlab/ferlease/cmd"
+import "github.com/Ferlab-Ste-Justine/ferlease/cmd"
 
 func main() {
 	cmd.Execute()
